fix: improving tests
