+++ packages/plugin/tests/processor-with-config.spec.ts
@@ -18,28 +18,15 @@ jest.mock('../src/graphql-config', () => ({
   })),
 }));
 
-describe('processor', () => {
+describe('processor with graphql-config', () => {
   const QUERY = 'query users { id }';
   const filePath = 'queries.ts';
-
-  it('preprocess finds gql tag', () => {
-    const code = `
-      import { gql } from 'graphql'
-      const fooQuery = gql\`${QUERY}\`
-    `;
-
-    const blocks = processor.preprocess(code, filePath);
-
-    expect(blocks[0].text).toBe(QUERY);
-    expect(blocks).toMatchSnapshot();
-  });
+  const code = `
+    import { custom } from 'custom-gql-tag'
+    const fooQuery = custom\`${QUERY}\`
+  `;
 
   it('preprocess finds custom tag', () => {
-    const code = `
-      import { custom } from 'custom-gql-tag'
-      const fooQuery = custom\`${QUERY}\`
-    `;
-
     const blocks = processor.preprocess(code, filePath);
 
     expect(blocks[0].text).toBe(QUERY);
